Retry after errors when watching a key.
It seems like the recent versions of etcd (or the etcd client) are returning
sporatic, spurious errors from the Watch command, but we really want to try to
avoid exiting the watch loop and closing the channels.  So we simply notify of
the errors and continue on. In the case of a severe failure (e.g. network is
down), this will eventually give up.