Etcd client lib: track errors per minute rather than total errors.
Previously, we kept a count of the absolute number of errors that had been
seen.  In load tests we are getting an error every 5 minutes, so this
eventually kills the server no matter what.

Instead, we now track errors-per-minute.  If there are more than 100
errors in a minute, we shut down.  This allows the slow, periodic errors
to continue forever as long as they appear to be recoverable.